cloudfunction/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
openai>=1.12.0
pyyaml>=6.0.1
tenacity>=8.2.3
python-multipart>=0.0.6
requests>=2.31.0
psutil>=5.9.8
aiohttp>=3.9.3
watchdog>=3.0.0 